move SelectedTab to it's own file
